fix(admin): put the server card where it can be seen
Caught by opening /admin on dev after deploying #758: the card was not there.
AdminComponent is not routed. /admin redirects to /admin/users, and the sidebar
links to users, webhooks and settings directly, so the landing page has been
dead code -- which is also why its four hardcoded English strings survived
every translation pass. Nobody could get to them.

The card moves to the top of Admin -> Settings, extracted as its own component
rather than more markup in a file that is already long, and testable on its own,
which the version on an unrouted page was not. The dead page and the five
translation keys that existed only for it are deleted.

Shipping an invisible control is the same defect this card was built to expose,
so: found in a browser, not in a test. The unit tests passed on the dead page
too.

Refs #730
