+++ Applications/Pgan.PoracleWebNet.App/ClientApp/src/app/modules/admin/admin-settings.component.html
@@ -6,6 +6,8 @@
 </div>
 
 <div class="page-content">
+  <app-server-profile-card />
+
   @if (settingsLoading()) {
     <div class="loading-container">
       <mat-spinner diameter="48"></mat-spinner>
