+++ Applications/Pgan.PoracleWebNet.App/ClientApp/src/app/modules/admin/admin-settings.component.ts
@@ -29,6 +29,7 @@ import { DiscordServerConfig, OidcServerConfig, PwebSetting, SiteSetting, Telegr
 import { I18nService } from '../../core/services/i18n.service';
 import { SettingsService } from '../../core/services/settings.service';
 import { ConfirmDialogComponent, ConfirmDialogData } from '../../shared/components/confirm-dialog/confirm-dialog.component';
+import { ServerProfileCardComponent } from '../../shared/components/server-profile-card/server-profile-card.component';
 
 /** Union type for backward compatibility during migration */
 type AnySettingItem = PwebSetting | SiteSetting;
@@ -335,6 +336,7 @@ const SETTING_GROUPS: SettingGroup[] = [
     MatSlideToggleModule,
     MatDividerModule,
     MatTooltipModule,
+    ServerProfileCardComponent,
     TranslatePipe,
   ],
   selector: 'app-admin-settings',
