+++ Applications/Pgan.PoracleWebNet.App/ClientApp/src/assets/i18n/da.json
@@ -1389,11 +1389,6 @@
     "SNACK_FAILED_REJECT": "Kunne ikke afvise indsendelse",
     "SNACK_REJECTED": "\"{{name}}\" afvist",
     "APPROVAL_REGION_HINT": "Vælg den region, dette geofence skal vises under.",
-    "PANEL_TITLE": "Administration",
-    "PANEL_DESC": "Vælg et område nedenfor.",
-    "NAV_USERS": "Brugere",
-    "NAV_WEBHOOKS": "Webhooks",
-    "NAV_SETTINGS": "Indstillinger",
     "SERVER_TITLE": "Poracle-server",
     "SERVER_REFRESH": "Tjek igen",
     "SERVER_VERSION": "Version",
