+++ Applications/Pgan.PoracleWebNet.App/ClientApp/src/assets/i18n/de.json
@@ -1389,11 +1389,6 @@
     "SNACK_FAILED_REJECT": "Einreichung konnte nicht abgelehnt werden",
     "SNACK_REJECTED": "\"{{name}}\" abgelehnt",
     "APPROVAL_REGION_HINT": "Wähle die Region, unter der dieses Geofence erscheinen soll.",
-    "PANEL_TITLE": "Administration",
-    "PANEL_DESC": "Wähle unten einen Bereich.",
-    "NAV_USERS": "Benutzer",
-    "NAV_WEBHOOKS": "Webhooks",
-    "NAV_SETTINGS": "Einstellungen",
     "SERVER_TITLE": "Poracle-Server",
     "SERVER_REFRESH": "Erneut prüfen",
     "SERVER_VERSION": "Version",
