+++ Applications/Pgan.PoracleWebNet.App/ClientApp/src/assets/i18n/en.json
@@ -1389,11 +1389,6 @@
     "SNACK_APPROVED": "\"{{name}}\" approved",
     "SNACK_FAILED_REJECT": "Failed to reject submission",
     "SNACK_REJECTED": "\"{{name}}\" rejected",
-    "PANEL_TITLE": "Admin Panel",
-    "PANEL_DESC": "Select a section below.",
-    "NAV_USERS": "Users",
-    "NAV_WEBHOOKS": "Webhooks",
-    "NAV_SETTINGS": "Settings",
     "SERVER_TITLE": "Poracle server",
     "SERVER_REFRESH": "Check again",
     "SERVER_VERSION": "Version",
