+++ Applications/Pgan.PoracleWebNet.App/ClientApp/src/assets/i18n/es.json
@@ -1389,11 +1389,6 @@
     "SNACK_FAILED_REJECT": "Error al rechazar el envío",
     "SNACK_REJECTED": "\"{{name}}\" rechazada",
     "APPROVAL_REGION_HINT": "Elige la región bajo la que aparecerá esta geocerca.",
-    "PANEL_TITLE": "Panel de administración",
-    "PANEL_DESC": "Elige una sección.",
-    "NAV_USERS": "Usuarios",
-    "NAV_WEBHOOKS": "Webhooks",
-    "NAV_SETTINGS": "Ajustes",
     "SERVER_TITLE": "Servidor Poracle",
     "SERVER_REFRESH": "Volver a comprobar",
     "SERVER_VERSION": "Versión",
