+++ Applications/Pgan.PoracleWebNet.App/ClientApp/src/assets/i18n/fr.json
@@ -1389,11 +1389,6 @@
     "SNACK_FAILED_REJECT": "Échec du rejet de la soumission",
     "SNACK_REJECTED": "\"{{name}}\" rejetée",
     "APPROVAL_REGION_HINT": "Choisissez la région sous laquelle ce géorepère apparaîtra.",
-    "PANEL_TITLE": "Panneau d’administration",
-    "PANEL_DESC": "Choisissez une section.",
-    "NAV_USERS": "Utilisateurs",
-    "NAV_WEBHOOKS": "Webhooks",
-    "NAV_SETTINGS": "Paramètres",
     "SERVER_TITLE": "Serveur Poracle",
     "SERVER_REFRESH": "Vérifier à nouveau",
     "SERVER_VERSION": "Version",
