+++ Applications/Pgan.PoracleWebNet.App/ClientApp/src/assets/i18n/it.json
@@ -1389,11 +1389,6 @@
     "SNACK_FAILED_REJECT": "Rifiuto invio fallito",
     "SNACK_REJECTED": "\"{{name}}\" rifiutata",
     "APPROVAL_REGION_HINT": "Scegli la regione sotto cui apparirà questo geofence.",
-    "PANEL_TITLE": "Pannello di amministrazione",
-    "PANEL_DESC": "Scegli una sezione.",
-    "NAV_USERS": "Utenti",
-    "NAV_WEBHOOKS": "Webhook",
-    "NAV_SETTINGS": "Impostazioni",
     "SERVER_TITLE": "Server Poracle",
     "SERVER_REFRESH": "Controlla di nuovo",
     "SERVER_VERSION": "Versione",
