+++ Applications/Pgan.PoracleWebNet.App/ClientApp/src/assets/i18n/nl.json
@@ -1389,11 +1389,6 @@
     "SNACK_FAILED_REJECT": "Inzending afwijzen mislukt",
     "SNACK_REJECTED": "\"{{name}}\" afgewezen",
     "APPROVAL_REGION_HINT": "Kies de regio waaronder deze geofence verschijnt.",
-    "PANEL_TITLE": "Beheerpaneel",
-    "PANEL_DESC": "Kies hieronder een onderdeel.",
-    "NAV_USERS": "Gebruikers",
-    "NAV_WEBHOOKS": "Webhooks",
-    "NAV_SETTINGS": "Instellingen",
     "SERVER_TITLE": "Poracle-server",
     "SERVER_REFRESH": "Opnieuw controleren",
     "SERVER_VERSION": "Versie",
