+++ Applications/Pgan.PoracleWebNet.App/ClientApp/src/assets/i18n/pl.json
@@ -1389,11 +1389,6 @@
     "SNACK_FAILED_REJECT": "Nie udało się odrzucić zgłoszenia",
     "SNACK_REJECTED": "\"{{name}}\" odrzucony",
     "APPROVAL_REGION_HINT": "Wybierz region, w którym pojawi się ten geofence.",
-    "PANEL_TITLE": "Panel administratora",
-    "PANEL_DESC": "Wybierz sekcję poniżej.",
-    "NAV_USERS": "Użytkownicy",
-    "NAV_WEBHOOKS": "Webhooki",
-    "NAV_SETTINGS": "Ustawienia",
     "SERVER_TITLE": "Serwer Poracle",
     "SERVER_REFRESH": "Sprawdź ponownie",
     "SERVER_VERSION": "Wersja",
