+++ Applications/Pgan.PoracleWebNet.App/ClientApp/src/assets/i18n/pt-BR.json
@@ -1389,11 +1389,6 @@
     "SNACK_FAILED_REJECT": "Falha ao rejeitar envio",
     "SNACK_REJECTED": "\"{{name}}\" rejeitada",
     "APPROVAL_REGION_HINT": "Escolha a região sob a qual esta geocerca aparecerá.",
-    "PANEL_TITLE": "Painel de administração",
-    "PANEL_DESC": "Escolha uma seção abaixo.",
-    "NAV_USERS": "Usuários",
-    "NAV_WEBHOOKS": "Webhooks",
-    "NAV_SETTINGS": "Configurações",
     "SERVER_TITLE": "Servidor Poracle",
     "SERVER_REFRESH": "Verificar de novo",
     "SERVER_VERSION": "Versão",
