+++ Applications/Pgan.PoracleWebNet.App/ClientApp/src/assets/i18n/pt.json
@@ -1389,11 +1389,6 @@
     "SNACK_FAILED_REJECT": "Falha ao rejeitar submissão",
     "SNACK_REJECTED": "\"{{name}}\" rejeitada",
     "APPROVAL_REGION_HINT": "Escolhe a região sob a qual esta geocerca aparecerá.",
-    "PANEL_TITLE": "Painel de administração",
-    "PANEL_DESC": "Escolhe uma secção abaixo.",
-    "NAV_USERS": "Utilizadores",
-    "NAV_WEBHOOKS": "Webhooks",
-    "NAV_SETTINGS": "Definições",
     "SERVER_TITLE": "Servidor Poracle",
     "SERVER_REFRESH": "Verificar de novo",
     "SERVER_VERSION": "Versão",
