+++ Applications/Pgan.PoracleWebNet.App/ClientApp/src/assets/i18n/sv.json
@@ -1389,11 +1389,6 @@
     "SNACK_FAILED_REJECT": "Kunde inte avvisa inskickning",
     "SNACK_REJECTED": "\"{{name}}\" avvisad",
     "APPROVAL_REGION_HINT": "Välj den region som detta geostängsel ska visas under.",
-    "PANEL_TITLE": "Administration",
-    "PANEL_DESC": "Välj ett avsnitt nedan.",
-    "NAV_USERS": "Användare",
-    "NAV_WEBHOOKS": "Webhooks",
-    "NAV_SETTINGS": "Inställningar",
     "SERVER_TITLE": "Poracle-server",
     "SERVER_REFRESH": "Kontrollera igen",
     "SERVER_VERSION": "Version",
