+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- **The Poracle server card is on a page you can actually open.** It shipped on the admin landing page, which nothing routes to — `/admin` redirects straight to Users, and the sidebar links to the three sections directly. The card now sits at the top of Admin → Settings, and the dead landing page is gone. Its four English-only strings had survived every translation pass for the same reason nobody noticed the page: you could not get to it ([#730](https://github.com/PGAN-Dev/PoracleWeb.NET/issues/730)).
 - **Max Battles and Fort Changes speak your language now.** Nine strings in those two sections were written straight into the templates and so appeared in English whatever the display language was set to — the hints explaining what each Max Battle tab tracks, what the Gigantamax toggle does, and the Fort Type and Change Types headings. All nine are translated into the ten other locales, matching the wording each already uses for "Max Battle" and "fort". The PVP level cap's screen-reader label was English too, and now reuses the translated legend above it ([#730](https://github.com/PGAN-Dev/PoracleWeb.NET/issues/730)).
 - **A quest's minimum amount can be changed after you set it.** It shipped in the add dialog only, so a card reading "3x Rare Candy" had no route back to the 3 — the same create-only shape as the mega picker, found by sweeping every add dialog against its edit twin. The stardust floor is editable there too. What stays fixed is which reward the alarm is about: changing Rare Candy to a Poke Ball describes a different alarm, not an edit of this one. A test now compares the two dialogs of all nine alarm types and fails on any control that exists in one and not the other, unless the difference is written down with a reason ([#730](https://github.com/PGAN-Dev/PoracleWeb.NET/issues/730)).
 - **The mega evolution picker is reachable when editing an alarm, not just when creating one.** In the edit dialog it sat inside the PVP level-cap fieldset, which renders only on a server that advertises level caps — so on a server without them a mega rule could be created and then never changed, and the card said "Mega X" with no way to get back to it. It is now a sibling of the cap picker, the way the add dialog has always had it ([#730](https://github.com/PGAN-Dev/PoracleWeb.NET/issues/730)).
